+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/angular
 
+## 3.0.43
+
+### Patch Changes
+
+- Updated dependencies [d8210b6]
+- Updated dependencies [b192878]
+  - @ai-sdk/provider-utils@5.0.16
+  - ai@7.0.43
+
 ## 3.0.42
 
 ### Patch Changes